requirements.txt
asgiref==3.5.0
Django==4.0.4
django-extensions==3.1.5
djangorestframework==3.13.1
flake8==4.0.1
mccabe==0.6.1
model-mommy==2.0.0
pycodestyle==2.8.0
pyflakes==2.4.0
pytz==2022.1
sqlparse==0.4.2
